Update utility testing to cover new altered utility functions - Migrate out all not utility testing
